+++ src/sections/stats/components/PieChart/PieChart.tsx
@@ -1,29 +1,19 @@
-import { Fragment, useMemo, useState } from "react"
-import { SClipPath, SLabelContainer, SSliceContainer } from "./PieChart.styled"
+import { ReactNode } from 'react'
 import {
   ASSET_COLORS,
-  getCircleCoordinates,
-  getPieConfig,
 } from "./PieChart.utils"
 import {
-  SliceLabelRest,
-  TLabelRest,
-} from "./components/SliceLabelRest/SliceLabelRest"
-import { AnimatePresence } from "framer-motion"
+  SliceLabelRest, TLabelRest,
+} from './components/SliceLabelRest/SliceLabelRest'
 import { SliceLabel } from "./components/SliceLabel/SliceLabel"
-import { DefaultSliceLabel } from "./components/DefaultSliceLabel/DefaultSliceLabel"
 import { PieSkeleton } from "./components/Skeleton/Skeleton"
-import { ScrollablePicker } from "../ScrollablePicker/ScrollablePicker"
-import { useMedia } from "react-use"
-import { theme } from "theme"
 import { t } from "i18next"
-import { EmotionJSX } from "@emotion/react/types/jsx-namespace"
-import { Text } from "components/Typography/Text/Text"
-import { useTranslation } from "react-i18next"
 import { TOmnipoolOverviewData } from "sections/stats/sections/overview/data/OmnipoolOverview.utils"
 import { BN_0 } from "utils/constants"
+import { DoughnutChart } from '../DoughnutChart/DoughnutChart'
+import { EmotionJSX } from '@emotion/react/types/jsx-namespace'
 
-export type TSlice = {
+type TSlice = {
   percentage: number
   color: string
   label: EmotionJSX.Element
@@ -32,141 +22,12 @@ export type TSlice = {
   assets?: TLabelRest[]
 }
 
-type DoughnutChartProps = {
-  slices: TSlice[]
-}
-
-const DoughnutChart = ({ slices }: DoughnutChartProps) => {
-  const isDesktop = useMedia(theme.viewport.gte.sm)
-  const PIE_SIZE = !isDesktop ? 170 : 300
-  const config = getPieConfig(PIE_SIZE)
-  const { t } = useTranslation()
-
-  const [activeSlice, setActiveSlice] = useState<number | undefined>(undefined)
-
-  const restCmp = slices.find((slice) => slice.symbol === "rest")
-  const restIsSelected = restCmp && slices.length - 1 === activeSlice
-
-  let label = <DefaultSliceLabel slices={slices} />
-
-  if (activeSlice != null) {
-    if (!isDesktop && restIsSelected) {
-      label = (
-        <Text color="basic100" fs={[20, 34]}>
-          {t("value.percentage", { value: restCmp.percentage })}
-        </Text>
-      )
-    } else {
-      label = slices[activeSlice].label
-    }
-  }
-
-  const sliceComponents = useMemo(() => {
-    if (!slices.length) return null
-
-    let diffAngle = 0
-    const components = slices.map((slice, index) => {
-      const sliceLength = slice.percentage * 3.6
-      const startAngle = diffAngle
-      const endAngle = sliceLength
-      diffAngle += endAngle
-
-      return (
-        <Fragment key={slice.percentage}>
-          <SClipPath
-            rotate={startAngle}
-            length={sliceLength - 1}
-            color={slice.color}
-            size={config.shadowSize}
-            radial
-            clipPath={getCircleCoordinates(
-              config.shadowInnerRadius,
-              config.innerRadius,
-              config.shadowSize,
-              slice.percentage,
-            )}
-          />
-
-          <SClipPath
-            rotate={startAngle}
-            length={sliceLength - 1}
-            color={slice.color}
-            size={config.pieSize}
-            clipPath={getCircleCoordinates(
-              config.innerRadius,
-              config.outerRadius,
-              config.pieSize,
-              slice.percentage,
-            )}
-          />
-
-          <SClipPath
-            rotate={startAngle}
-            length={sliceLength - 1}
-            color={slice.color}
-            size={config.pieSize}
-            isActive={activeSlice === index}
-            clipPath={getCircleCoordinates(
-              config.innerRadius,
-              config.outerRadius,
-              config.pieSize,
-              slice.percentage,
-            )}
-            hoverClipPath={getCircleCoordinates(
-              config.innerRadius,
-              config.hoverOuterRadius,
-              config.pieSize,
-              slice.percentage,
-            )}
-            onMouseMove={() => setActiveSlice(index)}
-            onMouseLeave={() => setActiveSlice(undefined)}
-          />
-        </Fragment>
-      )
-    })
-
-    return <SSliceContainer size={config.pieSize}>{components}</SSliceContainer>
-  }, [config, slices, activeSlice])
-
-  return (
-    <div sx={{ flex: "column", gap: 24 }}>
-      <div sx={{ flex: "row", justify: "space-between", align: "center" }}>
-        <div css={{ position: "relative" }}>
-          <SLabelContainer size={PIE_SIZE - config.shadowInnerRadius}>
-            <AnimatePresence>{label}</AnimatePresence>
-          </SLabelContainer>
-          <svg width={PIE_SIZE} height={PIE_SIZE}>
-            <foreignObject width={PIE_SIZE} height={PIE_SIZE}>
-              {sliceComponents}
-            </foreignObject>
-          </svg>
-        </div>
-        {!isDesktop ? (
-          <ScrollablePicker
-            values={[
-              {
-                name: t("stats.overview.pie.rest.label"),
-                color: "white",
-                symbol: "overview",
-                label: <Fragment />,
-                percentage: 0,
-              },
-              ...slices,
-            ]}
-            onChange={setActiveSlice}
-          />
-        ) : null}
-      </div>
-      {!isDesktop && restIsSelected ? restCmp.label : null}
-    </div>
-  )
-}
-
 type PieChartProps = {
   data: TOmnipoolOverviewData
+  sliceLabel?: (slices: TSlice[]) => ReactNode;
 }
 
-export const PieChart = ({ data }: PieChartProps) => {
+export const PieChart = ({ data, sliceLabel }: PieChartProps) => {
   if (!data) return <PieSkeleton />
 
   const tvlTotal = data.reduce(
@@ -175,7 +36,7 @@ export const PieChart = ({ data }: PieChartProps) => {
   )
 
   const slices = data
-    ?.reduce((acc, omnipoolAsset) => {
+    ?.reduce<TSlice[]>((acc, omnipoolAsset) => {
       const percentage = Number(
         omnipoolAsset.tvl.div(tvlTotal).multipliedBy(100).toFixed(2),
       )
@@ -244,8 +105,8 @@ export const PieChart = ({ data }: PieChartProps) => {
       }
 
       return acc
-    }, [] as TSlice[])
+    }, [])
     .sort((a) => (a.symbol !== "rest" ? -1 : 0))
 
-  return <DoughnutChart slices={slices} />
+  return <DoughnutChart slices={slices} sliceLabel={sliceLabel} />
 }
