+++ src/sections/stats/sections/LRNA/StatsLRNA.tsx
@@ -2,24 +2,42 @@ import { useApiPromise } from "utils/api"
 import { ChartWrapper } from "./components/ChartWrapper/ChartWrapper"
 import { PieWrapper } from "./components/PieWrapper/PieWrapper"
 import { isApiLoaded } from "utils/helpers"
-import { useOmnipoolOverviewData } from "./data/OmnipoolOverview.utils"
 import { useMedia } from "react-use"
 import { theme } from "theme"
 import { SContainerVertical } from "./StatsOverview.styled"
 import { RecentTradesTableWrapper } from "./components/RecentTradesTable/RecentTradesTableWrapper"
 import { StatsTiles } from './components/StatsTiles/StatsTiles'
 import { PieChart } from './components/PieChart/PieChart'
+import BN from 'bignumber.js'
 
 const StatsLRNAData = () => {
   const isDesktop = useMedia(theme.viewport.gte.sm)
-  const omnipoolOverview = useOmnipoolOverviewData()
 
   return (
     <div sx={{ flex: "column", gap: 50 }}>
       <div sx={{ flex: "row", gap: 20 }}>
         <PieWrapper
-          data={omnipoolOverview.data}
-          isLoading={omnipoolOverview.isLoading}
+          data={[
+            {
+              id: "0",
+              name: "HDX",
+              symbol: "HDX",
+              tvl: new BN(1234),
+              volume: new BN(0),
+              fee: new BN(0),
+              pol: new BN(567)
+            },
+            {
+              id: "1",
+              name: "Ether (via Wormhole)",
+              symbol: "ETH",
+              tvl: new BN(890),
+              volume: new BN(0),
+              fee: new BN(0),
+              pol: new BN(123)
+            }
+          ]}
+          isLoading={false}
         />
         {isDesktop && (
           <SContainerVertical
@@ -32,8 +50,9 @@ const StatsLRNAData = () => {
           </SContainerVertical>
         )}
       </div>
-
-      <PieChart percentage={20} loading={false} />
+      <SContainerVertical sx={{ width: "100%", p: 40 }}>
+        <PieChart percentage={20} loading={false} />
+      </SContainerVertical>
       <StatsTiles />
       <RecentTradesTableWrapper />
     </div>
